Refactor TLS setup code into a separate file and consolidate logic between the StatefulSet and exporter Deployment
